move e2e framework.ProviderInterface to framework/providers package
Part of the refactoring to fix #109391

A note about the removal of `FrameworkBeforeEach(f *Framework)` and `FrameworkAfterEach(f *Framework)` from `ProviderInterface`:

These two methods reference `framework.Framework` directly, which is what we are trying to avoid, and currently none of these two methods are used or implemented, so it's safe to delete them.
